+++ autoware_utils/package.xml
@@ -19,12 +19,12 @@
   <depend>autoware_internal_planning_msgs</depend>
   <depend>autoware_perception_msgs</depend>
   <depend>autoware_planning_msgs</depend>
-  <depend>autoware_utils_geometry</depend>
   <depend>autoware_utils_logging</depend>
   <depend>autoware_utils_math</depend>
   <depend>autoware_utils_pcl</depend>
   <depend>autoware_utils_rclcpp</depend>
   <depend>autoware_utils_system</depend>
+  <depend>autoware_utils_tf</depend>
   <depend>autoware_utils_uuid</depend>
   <depend>autoware_utils_visualization</depend>
   <depend>autoware_vehicle_msgs</depend>
