+++ autoware_utils_pcl/package.xml
@@ -14,7 +14,7 @@
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
   <buildtool_depend>autoware_cmake</buildtool_depend>
 
-  <depend>autoware_utils_geometry</depend>
+  <depend>autoware_utils_tf</depend>
   <depend>pcl_conversions</depend>
   <depend>pcl_ros</depend>
