+++ autoware_utils_tf/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.14)
-project(autoware_utils_geometry)
+project(autoware_utils_tf)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
